Moves all global enums to a central file
