+++ Marlin/MarlinSerial.cpp
@@ -455,21 +455,6 @@ MarlinSerial customizedSerial;
 
   FORCE_INLINE void emergency_parser(unsigned char c) {
 
-    enum e_parser_state {
-      state_RESET,
-      state_N,
-      state_M,
-      state_M1,
-      state_M10,
-      state_M108,
-      state_M11,
-      state_M112,
-      state_M4,
-      state_M41,
-      state_M410,
-      state_IGNORE // to '\n'
-    };
-
     static e_parser_state state = state_RESET;
 
     switch (state) {
