+++ Marlin/cardreader.h
@@ -28,7 +28,7 @@
 #define MAX_DIR_DEPTH 10          // Maximum folder depth
 
 #include "SdFile.h"
-enum LsAction { LS_SerialPrint, LS_Count, LS_GetFilename };
+#include "enum.h"
 
 class CardReader {
 public:
