+++ Marlin/endstops.h
@@ -27,7 +27,7 @@
 #ifndef ENDSTOPS_H
 #define ENDSTOPS_H
 
-enum EndstopEnum {X_MIN = 0, Y_MIN = 1, Z_MIN = 2, Z_MIN_PROBE = 3, X_MAX = 4, Y_MAX = 5, Z_MAX = 6, Z2_MIN = 7, Z2_MAX = 8};
+#include "enum.h"
 
 class Endstops {
 
@@ -42,7 +42,7 @@ class Endstops {
       static byte
     #endif
         current_endstop_bits, old_endstop_bits;
-        
+
     Endstops() {};
 
     /**
