+++ Marlin/ultralcd.h
@@ -73,16 +73,6 @@
     void lcd_ignore_click(bool b=true);
 
     #if ENABLED(FILAMENT_CHANGE_FEATURE)
-      enum FilamentChangeMessage {
-        FILAMENT_CHANGE_MESSAGE_INIT,
-        FILAMENT_CHANGE_MESSAGE_UNLOAD,
-        FILAMENT_CHANGE_MESSAGE_INSERT,
-        FILAMENT_CHANGE_MESSAGE_LOAD,
-        FILAMENT_CHANGE_MESSAGE_EXTRUDE,
-        FILAMENT_CHANGE_MESSAGE_OPTION,
-        FILAMENT_CHANGE_MESSAGE_RESUME,
-        FILAMENT_CHANGE_MESSAGE_STATUS
-      };
       void lcd_filament_change_show_message(FilamentChangeMessage message);
     #endif // FILAMENT_CHANGE_FEATURE
